requirements.txt
# Core dependencies
requests

# Discord bot
discord.py

# Slack bot
slack-bolt
slack-sdk

# Web chat (FastAPI + auth + templates)
fastapi
uvicorn
python-jose[cryptography]  
jinja2                      
python-multipart 

# Dev extras
black
flake8
pytest